pyproject.toml
[tool.poetry]
name = "yamllint-demo"
version = "0.1.0"
description = ""
authors = ["ges <gerrit.schrader@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
yamllint = "^1.25.0"
"ruamel.yaml" = "^0.16.12"
lark-parser = "^0.10.1"
schema = "^0.7.3"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
